# docs/swagger.yaml
definitions:
  main.Ad:
    properties:
      conditions:
        $ref: '#/definitions/main.Conditions'
      endAt:
        type: string
      startAt:
        type: string
      title:
        type: string
    type: object
  main.Conditions:
    properties:
      ageEnd:
        type: integer
      ageStart:
        type: integer
      country:
        items:
          type: string
        type: array
      gender:
        type: string
      platform:
        items:
          type: string
        type: array
    type: object
  main.QueryCondition:
    properties:
      age:
        type: integer
      country:
        type: string
      gender:
        type: string
      limit:
        type: integer
      offset:
        type: integer
      platform:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/ad:
    get:
      consumes:
      - application/json
      description: Retieve ad by Query
      parameters:
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Limit for pagination defalut 5
        in: query
        name: limit
        type: integer
      - description: Age to Query
        in: query
        name: age
        type: integer
      - description: Gender
        in: query
        name: gender
        type: string
      - description: Country
        in: query
        name: country
        type: string
      - description: Platform
        in: query
        name: platform
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.QueryCondition'
      summary: Retieve Ad
      tags:
      - Ads
    post:
      consumes:
      - application/json
      description: create ad by JSON
      parameters:
      - description: Create ad
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/main.Ad'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Ad'
      summary: Create Ad
      tags:
      - Ads
swagger: "2.0"